Better organized manual test for type collector.
