+++ code-gen/src/test/java/org/paninij/apt/util/VisitorDriver.java
@@ -16,7 +16,7 @@
  *
  * Contributor(s): David Johnston
  */
-package org.paninij.apt.test;
+package org.paninij.apt.util;
 
 import java.util.Set;
